# src/main/java/com/CustomerMs/CustomerMs/Model/api/customer/ClienteResponse.java
package com.CustomerMs.CustomerMs.Model.api.customer;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ClienteResponse {

    private Long id;
    private String nombre;
    private String apellido;
    private String dni;
    private String email;
}