# src/main/resources/OpenApi/Cliente.yml
openapi: 3.0.3
info:
  title: API de Clientes
  description: Microservicio para gestionar clientes.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
  /clientes:
    post:
      summary: Crear cliente
      description: Crea un nuevo cliente en el sistema
      tags:
        - Clientes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequest'
      responses:
        "200":
          description: Cliente creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCliente'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /clientes/{id}:
    get:
      summary: Obtener cliente por ID
      description: Obtiene la información de un cliente registrado
      tags:
        - Clientes
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCliente'
        "404":
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ClienteRequest:
      type: object
      properties:
        nombre:
          type: string
          example: Juan
        apellido:
          type: string
          example: Pérez
        dni:
          type: string
          example: "12345678"
        email:
          type: string
          example: juan.perez@example.com
      required:
        - nombre
        - apellido
        - dni
        - email

    ClienteResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Juan
        apellido:
          type: string
          example: Pérez
        dni:
          type: string
          example: "12345678"
        email:
          type: string
          example: juan.perez@example.com

    ResponseCliente:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Cliente creado con éxito
        data:
          $ref: '#/components/schemas/ClienteResponse'

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error al procesar la solicitud